split the media auto-switch into focused modules
media_auto_switch.py had grown to 1447 lines covering name resolution, locality
verification, backend probing, lock accounting, refusal shapes and the switch itself.
It is now the orchestration only, with each of those beside it: media_model_index,
media_locality, media_switch_backends, media_switch_locks and media_switch_errors.

Behaviour is unchanged. The locality refusal was written out twice and the busy
refusal three times, so both collapse into one helper; the pre-index resident
shortcut and the lock acquisition become named helpers; one dead import goes.

Two defects fixed on the way. A lone unlabelled GGUF was marked ambiguous because its
published quant token is empty, so every request naming it reloaded the pipeline that
was already resident; only a token two builds share is ambiguous. The handed-over setup
task's exception was never retrieved once the budget expired, so an ordinary refusal
from it was reported by the loop as an unhandled task exception.
